+++ source/source_estate/param_update.h
@@ -1,0 +1,15 @@
+#ifndef PARAM_UPDATE_H
+#define PARAM_UPDATE_H
+
+#include "source_cell/cal_atoms_info.h"
+
+namespace elecstate {
+
+class ParamUpdater {
+public:
+    static void update_from_atoms_info(const AtomsInfoResult& atoms_info);
+};
+
+}
+
+#endif
